requirements.txt
pyrogram==1.2.8
telethon==1.23.0
pymongo==3.12.0
cryptography==35.0.0
python-dotenv==0.19.1
python-telegram-bot

